crypto_fetcher/requirements.txt
requests==2.31.0
click==8.1.7
python-dateutil==2.8.2
pydantic==2.5.2
tqdm==4.66.1
python-crontab==3.0.0
